+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- [Breaking] Jakarta annotations and validation groups support [#472](https://github.com/leangen/graphql-spqr/pull/472)
+
 ## [0.12.3] - 2023-06-18
 
 ### Added